+++ app/src/main/java/com/sanmed/android/messageexpenses/view/utilities/CurrencyUtilities.kt
@@ -1,11 +1,12 @@
 package com.sanmed.android.messageexpenses.view.utilities
 
+import java.math.BigDecimal
 import java.text.NumberFormat
 
 object CurrencyUtilities {
           private val currencyFormat = NumberFormat.getCurrencyInstance()
 
-        fun format(amount:Float?):String{
+        fun format(amount:BigDecimal?):String{
             return currencyFormat.format(amount)
         }
